% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuplemapr-package.R
\docType{package}
\name{tuplemapr-package}
\alias{tuplemapr}
\alias{tuplemapr-package}
\title{tuplemapr: C++ functions for working with compile-time fixed-length data}
\description{
This package provides C++ functions for manipulating compile-time fixed-length data, such as tuples and arrays. The core is tuple::map, which accepts an invokable and any number of tuples as arguments. It return a tuple containing the application of the invokable on the first elements, second elements, etc.
}
\keyword{internal}
